Fix DownloadZone Move examples
